requirements.txt
openai==0.28.1
requests
python-dotenv
diffusers== 0.33.1
transformers==4.51.3
torch==2.7.0+cu118
Pillow==9.4.0